Finish a good pass over code reorganization. Now all packages are in internal/pkg as it should be
