+++ Makefile
@@ -3,10 +3,6 @@
 # LICENSE.md file distributed with the sources of this project regarding your
 # rights to use or distribute this software.
 
-SUBDIRS = configparser experiments results
-SOURCES = main.go
-TOPDIR := $(PWD)
-
 all:
 	go build main.go
