+++ internal/pkg/checker/distro.go
